fix(jvm): route runtime metrics through shared events
Emit JVM heap summary and memory-pool GC metrics through the shared events ring buffer with dedicated JVM event type IDs. Remove the dedicated JVM runtime metric ring buffers and dispatch JVM records from the shared parser path used by generic and Go tracers.

Move the JVM and USDT BPF map definitions into map headers so the BPF layout follows the existing map organization. Remove the JVMRuntimeEventTracer interface plumbing and store the JVM runtime metrics queue on the shared eBPF event context.

Address CI lint feedback by marking the BPF_UPROBE ctx parameter as intentionally unused and removing the redundant uint64 conversion in the USDT test.

Validation: go test ./pkg/ebpf/common ./pkg/appolly/app/runtime ./pkg/ebpf; golangci-lint run ./pkg/ebpf; focused clang-tidy misc-unused-parameters for bpf/generictracer/jvm.c; BPF x86 and arm64 clang compiles; Linux generictracer and gotracer test compiles; make compile; git diff --check.

+++ pkg/ebpf/usdt_test.go
@@ -55,7 +55,7 @@ func TestParseUSDTArgSpecX8664(t *testing.T) {
 
 	assert.Equal(t, obiUSDTArgRegDeref, spec.Args[2].ArgType)
 	assert.Equal(t, int16(152), spec.Args[2].RegOff)
-	assert.Equal(t, uint64(^uint64(15)), spec.Args[2].ValOff)
+	assert.Equal(t, ^uint64(15), spec.Args[2].ValOff)
 
 	assert.Equal(t, obiUSDTArgConst, spec.Args[3].ArgType)
 	assert.Equal(t, uint64(7), spec.Args[3].ValOff)
